local

+++ src/rotationNaturalImages.py
@@ -112,12 +112,10 @@ def train(args):
 
         if epoch_val_accuracy > best_val_acc:
             best_val_acc = epoch_val_accuracy
-            torch.save(model.state_dict(), f'/raid/ferles/checkpoints/eb0/{dataset}/rot_{training_configurations.checkpoint}.pth')
-
-            if best_val_acc - checkpoint_val_accuracy > 0.05:
-                checkpoint_val_accuracy = best_val_acc
-                torch.save(model.state_dict(), f'/raid/ferles/checkpoints/eb0/{dataset}/rot_{training_configurations.checkpoint}_epoch_{epoch}_accuracy_{best_val_acc}.pth')
-
+            if os.path.exists('/raid/ferles/'):
+                torch.save(model.state_dict(), f'/raid/ferles/checkpoints/eb0/{dataset}/rot_{training_configurations.checkpoint}.pth')
+            else:
+                torch.save(model.state_dict(), f'/home/ferles/checkpoints/eb0/{dataset}/rot_{training_configurations.checkpoint}.pth')
             correct, total = 0, 0
 
             for data in testloader:
